Refactor: Separate execute and error handling into dedicated files
