+++ src/ft_error.c
@@ -1,7 +1,7 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils_helper2.c                                    :+:      :+:    :+:   */
+/*   ft_error.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mkazuhik <mkazuhik@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
@@ -11,31 +11,9 @@
 /* ************************************************************************** */
 
 #include "../pipex.h"
+#include <stdio.h>
 
-char	*find_path(char *cmd, char **envp)
+void	ft_error(const char *msg)
 {
-	char	**paths;
-	char	*path;
-	int		i;
-
-	if (!cmd || !envp)
-		return (0);
-	i = 0;
-	while (ft_strnstr(envp[i], "PATH", 4) == 0)
-		i++;
-	paths = ft_split(envp[i] + 5, ':');
-	i = 0;
-	while (paths[i])
-	{
-		path = join_path(paths[i], cmd);
-		if (access(path, F_OK) == 0)
-		{
-			free_paths(paths);
-			return (path);
-		}
-		free(path);
-		i++;
-	}
-	free_paths(paths);
-	return (0);
+	perror(msg);
 }
